requirements.txt
fastapi
uvicorn
jinja2
pymongo
python-dotenv
insightface
onnxruntime
numpy
opencv-python-headless
